pyproject.toml
[project]
name = "reply-mirror-fraud"
version = "0.1.0"
description = "Agent-based fraud detection pipeline for Reply Mirror challenge"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pandas>=2.2.0",
  "numpy>=1.26.0",
  "scikit-learn>=1.4.0",
  "pydantic>=2.6.0",
  "typer>=0.12.0",
  "pyarrow>=15.0.0",
  "networkx>=3.2.1",
  "PyYAML>=6.0.1",
  "requests>=2.31.0",
]

[project.optional-dependencies]
audio = ["openai-whisper>=20231117"]
observability = ["langfuse>=2.50.0"]
dev = ["pytest>=8.1.0", "ruff>=0.4.0", "mypy>=1.9.0"]

[project.scripts]
mirror = "mirror.cli:app"

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["mirror*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
